Handle invalid history match in ggtags-global-filter
by jumping to first error instead.

For example when a history match is set on the second match of all
matches and the search later turns up only single match, ggtags will
fail to jump to match. To make things worse ggtags-global-handle-exit
will kill the *ggtags-global* buffer leaving users with no opportunity
to move to a valid match.