Allow `ggtags-global-output` to run the process synchronously
This will be useful when developing more frontends to ggtags which may need to
receive a list of tag matches synchronously, such as XREF. Alternatively, a
separate function which calls a process synchronously using something like
`process-file` could be written, but then the ability to cancel the subprocess
after a certain number of lines have been read would be lost. This is
important, for example, when asking Global for all references to a tag which
could have thousands of matches. Packages such as XREF cannot handle such a
large number of matches efficiently at the moment.